backend/requirements.txt
fastapi
uvicorn
sqlmodel
spacy
newspaper3k
python-docx
requests